Refactor: rename payment_method to payment_methods in take offer wizard
