+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/take_offer/TakeOfferController.java
@@ -25,7 +25,7 @@ import bisq.desktop.common.view.InitWithDataController;
 import bisq.desktop.common.view.Navigation;
 import bisq.desktop.common.view.NavigationController;
 import bisq.desktop.main.content.bisq_easy.take_offer.amount.TakeOfferAmountController;
-import bisq.desktop.main.content.bisq_easy.take_offer.payment_method.TakeOfferPaymentController;
+import bisq.desktop.main.content.bisq_easy.take_offer.payment_methods.TakeOfferPaymentController;
 import bisq.desktop.main.content.bisq_easy.take_offer.review.TakeOfferReviewController;
 import bisq.desktop.overlay.OverlayController;
 import bisq.i18n.Res;
