+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/take_offer/payment_methods/TakeOfferPaymentModel.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.bisq_easy.take_offer.payment_method;
+package bisq.desktop.main.content.bisq_easy.take_offer.payment_methods;
 
 import bisq.common.currency.Market;
 import bisq.desktop.common.view.Model;
